Move selectLatestSample from capacity-band.ts to its natural home usage-samples.ts, removing the cross-rendering-module dependency (the pace-position-panel -> capacity-band coupling flagged in #1404).
