+++ office-hours/src/capacity-band.ts
@@ -4,17 +4,6 @@ const MINUTE = 60_000;
 const HOUR = 3_600_000;
 const DAY = 86_400_000;
 
-/**
- * Returns the sample with the maximum sampledAt, or null for an empty array.
- * Does not mutate the input array.
- */
-export function selectLatestSample(samples: UsageSample[]): UsageSample | null {
-  if (samples.length === 0) return null;
-  return samples.reduce((best, s) =>
-    s.sampledAt.getTime() > best.sampledAt.getTime() ? s : best,
-  );
-}
-
 function workerState(active: number, target: number): string {
   if (target === 0) return "paused";
   if (active < target) return "spawning";
